Split X1 code
