+++ db/pgstorage/pgstorage.go
@@ -7,9 +7,7 @@ import (
 	"math/big"
 	"time"
 
-	"github.com/0xPolygonHermez/zkevm-bridge-service/bridgectrl/pb"
 	ctmtypes "github.com/0xPolygonHermez/zkevm-bridge-service/claimtxman/types"
-	"github.com/0xPolygonHermez/zkevm-bridge-service/config/apolloconfig"
 	"github.com/0xPolygonHermez/zkevm-bridge-service/etherman"
 	"github.com/0xPolygonHermez/zkevm-bridge-service/utils/gerror"
 	"github.com/0xPolygonHermez/zkevm-node/log"
@@ -22,7 +20,6 @@ import (
 // PostgresStorage implements the Storage interface.
 type PostgresStorage struct {
 	*pgxpool.Pool
-	tableSuffix apolloconfig.Entry[string]
 }
 
 // getExecQuerier determines which execQuerier to use, dbTx or the main pgxpool
@@ -45,7 +42,7 @@ func NewPostgresStorage(cfg Config) (*PostgresStorage, error) {
 		log.Errorf("Unable to connect to database: %v\n", err)
 		return nil, err
 	}
-	return &PostgresStorage{Pool: db, tableSuffix: apolloconfig.NewStringEntry("DB.TableSuffix", cfg.TableSuffix)}, nil
+	return &PostgresStorage{db}, nil
 }
 
 // Rollback rollbacks a db transaction.
@@ -73,7 +70,7 @@ func (p *PostgresStorage) BeginDBTransaction(ctx context.Context) (pgx.Tx, error
 // GetLastBlock gets the last block.
 func (p *PostgresStorage) GetLastBlock(ctx context.Context, networkID uint, dbTx pgx.Tx) (*etherman.Block, error) {
 	var block etherman.Block
-	getLastBlockSQL := fmt.Sprintf("SELECT id, block_num, block_hash, parent_hash, network_id, received_at FROM sync.block%[1]v where network_id = $1 ORDER BY block_num DESC LIMIT 1", p.tableSuffix)
+	const getLastBlockSQL = "SELECT id, block_num, block_hash, parent_hash, network_id, received_at FROM sync.block where network_id = $1 ORDER BY block_num DESC LIMIT 1"
 
 	e := p.getExecQuerier(dbTx)
 	err := e.QueryRow(ctx, getLastBlockSQL, networkID).Scan(&block.ID, &block.BlockNumber, &block.BlockHash, &block.ParentHash, &block.NetworkID, &block.ReceivedAt)
@@ -88,12 +85,12 @@ func (p *PostgresStorage) GetLastBlock(ctx context.Context, networkID uint, dbTx
 // AddBlock adds a new block to the storage.
 func (p *PostgresStorage) AddBlock(ctx context.Context, block *etherman.Block, dbTx pgx.Tx) (uint64, error) {
 	var blockID uint64
-	addBlockSQL := fmt.Sprintf(`WITH block_id AS 
-		(INSERT INTO sync.block%[1]v (block_num, block_hash, parent_hash, network_id, received_at) 
+	const addBlockSQL = `WITH block_id AS 
+		(INSERT INTO sync.block (block_num, block_hash, parent_hash, network_id, received_at) 
 		VALUES ($1, $2, $3, $4, $5) ON CONFLICT (block_hash) DO NOTHING RETURNING id)
 		SELECT * from block_id
 		UNION ALL
-		SELECT id FROM sync.block%[1]v WHERE block_hash = $2;`, p.tableSuffix)
+		SELECT id FROM sync.block WHERE block_hash = $2;`
 	e := p.getExecQuerier(dbTx)
 	err := e.QueryRow(ctx, addBlockSQL, block.BlockNumber, block.BlockHash, block.ParentHash, block.NetworkID, block.ReceivedAt).Scan(&blockID)
 
@@ -106,7 +103,7 @@ func (p *PostgresStorage) AddBlock(ctx context.Context, block *etherman.Block, d
 
 // AddGlobalExitRoot adds a new ExitRoot to the db.
 func (p *PostgresStorage) AddGlobalExitRoot(ctx context.Context, exitRoot *etherman.GlobalExitRoot, dbTx pgx.Tx) error {
-	addExitRootSQL := fmt.Sprintf("INSERT INTO sync.exit_root%[1]v (block_id, global_exit_root, exit_roots) VALUES ($1, $2, $3)", p.tableSuffix)
+	const addExitRootSQL = "INSERT INTO sync.exit_root (block_id, global_exit_root, exit_roots) VALUES ($1, $2, $3)"
 	e := p.getExecQuerier(dbTx)
 	_, err := e.Exec(ctx, addExitRootSQL, exitRoot.BlockID, exitRoot.GlobalExitRoot, pq.Array([][]byte{exitRoot.ExitRoots[0][:], exitRoot.ExitRoots[1][:]}))
 	return err
@@ -114,19 +111,16 @@ func (p *PostgresStorage) AddGlobalExitRoot(ctx context.Context, exitRoot *ether
 
 // AddDeposit adds new deposit to the storage.
 func (p *PostgresStorage) AddDeposit(ctx context.Context, deposit *etherman.Deposit, dbTx pgx.Tx) (uint64, error) {
-	addDepositSQL := fmt.Sprintf(`
-		INSERT INTO sync.deposit%[1]v (leaf_type, network_id, orig_net, orig_addr, amount, dest_net, dest_addr, block_id, deposit_cnt, tx_hash, metadata)
-		VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11) RETURNING id`, p.tableSuffix)
+	const addDepositSQL = "INSERT INTO sync.deposit (leaf_type, network_id, orig_net, orig_addr, amount, dest_net, dest_addr, block_id, deposit_cnt, tx_hash, metadata) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11) RETURNING id"
 	e := p.getExecQuerier(dbTx)
 	var depositID uint64
-	err := e.QueryRow(ctx, addDepositSQL, deposit.LeafType, deposit.NetworkID, deposit.OriginalNetwork, deposit.OriginalAddress, deposit.Amount.String(),
-		deposit.DestinationNetwork, deposit.DestinationAddress, deposit.BlockID, deposit.DepositCount, deposit.TxHash, deposit.Metadata).Scan(&depositID)
+	err := e.QueryRow(ctx, addDepositSQL, deposit.LeafType, deposit.NetworkID, deposit.OriginalNetwork, deposit.OriginalAddress, deposit.Amount.String(), deposit.DestinationNetwork, deposit.DestinationAddress, deposit.BlockID, deposit.DepositCount, deposit.TxHash, deposit.Metadata).Scan(&depositID)
 	return depositID, err
 }
 
 // AddClaim adds new claim to the storage.
 func (p *PostgresStorage) AddClaim(ctx context.Context, claim *etherman.Claim, dbTx pgx.Tx) error {
-	addClaimSQL := fmt.Sprintf("INSERT INTO sync.claim%[1]v (network_id, index, orig_net, orig_addr, amount, dest_addr, block_id, tx_hash) VALUES ($1, $2, $3, $4, $5, $6, $7, $8)", p.tableSuffix)
+	const addClaimSQL = "INSERT INTO sync.claim (network_id, index, orig_net, orig_addr, amount, dest_addr, block_id, tx_hash) VALUES ($1, $2, $3, $4, $5, $6, $7, $8)"
 	e := p.getExecQuerier(dbTx)
 	_, err := e.Exec(ctx, addClaimSQL, claim.NetworkID, claim.Index, claim.OriginalNetwork, claim.OriginalAddress, claim.Amount.String(), claim.DestinationAddress, claim.BlockID, claim.TxHash)
 	return err
@@ -135,7 +129,7 @@ func (p *PostgresStorage) AddClaim(ctx context.Context, claim *etherman.Claim, d
 // GetTokenMetadata gets the metadata of the dedicated token.
 func (p *PostgresStorage) GetTokenMetadata(ctx context.Context, networkID, destNet uint, originalTokenAddr common.Address, dbTx pgx.Tx) ([]byte, error) {
 	var metadata []byte
-	getMetadataSQL := fmt.Sprintf("SELECT metadata from sync.deposit%[1]v WHERE network_id = $1 AND orig_addr = $2 AND dest_net = $3 AND metadata IS NOT NULL LIMIT 1", p.tableSuffix)
+	const getMetadataSQL = "SELECT metadata from sync.deposit WHERE network_id = $1 AND orig_addr = $2 AND dest_net = $3 AND metadata IS NOT NULL LIMIT 1"
 	e := p.getExecQuerier(dbTx)
 	err := e.QueryRow(ctx, getMetadataSQL, networkID, originalTokenAddr, destNet).Scan(&metadata)
 	return metadata, err
@@ -159,18 +153,15 @@ func (p *PostgresStorage) AddTokenWrapped(ctx context.Context, tokenWrapped *eth
 		}
 	}
 
-	addTokenWrappedSQL := fmt.Sprintf(`
-		INSERT INTO sync.token_wrapped%[1]v (network_id, orig_net, orig_token_addr, wrapped_token_addr, block_id, name, symbol, decimals)
-		VALUES ($1, $2, $3, $4, $5, $6, $7, $8)`, p.tableSuffix)
+	const addTokenWrappedSQL = "INSERT INTO sync.token_wrapped (network_id, orig_net, orig_token_addr, wrapped_token_addr, block_id, name, symbol, decimals) VALUES ($1, $2, $3, $4, $5, $6, $7, $8)"
 	e := p.getExecQuerier(dbTx)
-	_, err = e.Exec(ctx, addTokenWrappedSQL, tokenWrapped.NetworkID, tokenWrapped.OriginalNetwork, tokenWrapped.OriginalTokenAddress,
-		tokenWrapped.WrappedTokenAddress, tokenWrapped.BlockID, tokenMetadata.Name, tokenMetadata.Symbol, tokenMetadata.Decimals)
+	_, err = e.Exec(ctx, addTokenWrappedSQL, tokenWrapped.NetworkID, tokenWrapped.OriginalNetwork, tokenWrapped.OriginalTokenAddress, tokenWrapped.WrappedTokenAddress, tokenWrapped.BlockID, tokenMetadata.Name, tokenMetadata.Symbol, tokenMetadata.Decimals)
 	return err
 }
 
 // Reset resets the state to a block for the given DB tx.
 func (p *PostgresStorage) Reset(ctx context.Context, blockNumber uint64, networkID uint, dbTx pgx.Tx) error {
-	resetSQL := fmt.Sprintf("DELETE FROM sync.block%[1]v WHERE block_num > $1 AND network_id = $2", p.tableSuffix)
+	const resetSQL = "DELETE FROM sync.block WHERE block_num > $1 AND network_id = $2"
 	e := p.getExecQuerier(dbTx)
 	_, err := e.Exec(ctx, resetSQL, blockNumber, networkID)
 	return err
@@ -179,7 +170,7 @@ func (p *PostgresStorage) Reset(ctx context.Context, blockNumber uint64, network
 // GetPreviousBlock gets the offset previous L1 block respect to latest.
 func (p *PostgresStorage) GetPreviousBlock(ctx context.Context, networkID uint, offset uint64, dbTx pgx.Tx) (*etherman.Block, error) {
 	var block etherman.Block
-	getPreviousBlockSQL := fmt.Sprintf("SELECT block_num, block_hash, parent_hash, network_id, received_at FROM sync.block%[1]v WHERE network_id = $1 ORDER BY block_num DESC LIMIT 1 OFFSET $2", p.tableSuffix)
+	const getPreviousBlockSQL = "SELECT block_num, block_hash, parent_hash, network_id, received_at FROM sync.block WHERE network_id = $1 ORDER BY block_num DESC LIMIT 1 OFFSET $2"
 	e := p.getExecQuerier(dbTx)
 	err := e.QueryRow(ctx, getPreviousBlockSQL, networkID, offset).Scan(&block.BlockNumber, &block.BlockHash, &block.ParentHash, &block.NetworkID, &block.ReceivedAt)
 	if errors.Is(err, pgx.ErrNoRows) {
@@ -191,17 +182,17 @@ func (p *PostgresStorage) GetPreviousBlock(ctx context.Context, networkID uint, 
 // GetNumberDeposits gets the number of  deposits.
 func (p *PostgresStorage) GetNumberDeposits(ctx context.Context, networkID uint, blockNumber uint64, dbTx pgx.Tx) (uint64, error) {
 	var nDeposits int64
-	getNumDepositsSQL := fmt.Sprintf("SELECT coalesce(MAX(deposit_cnt), -1) FROM sync.deposit%[1]v as d INNER JOIN sync.block%[1]v as b ON d.network_id = b.network_id AND d.block_id = b.id WHERE d.network_id = $1 AND b.block_num <= $2", p.tableSuffix)
+	const getNumDepositsSQL = "SELECT coalesce(MAX(deposit_cnt), -1) FROM sync.deposit as d INNER JOIN sync.block as b ON d.network_id = b.network_id AND d.block_id = b.id WHERE d.network_id = $1 AND b.block_num <= $2"
 	err := p.getExecQuerier(dbTx).QueryRow(ctx, getNumDepositsSQL, networkID, blockNumber).Scan(&nDeposits)
 	return uint64(nDeposits + 1), err
 }
 
 // AddTrustedGlobalExitRoot adds new global exit root which comes from the trusted sequencer.
 func (p *PostgresStorage) AddTrustedGlobalExitRoot(ctx context.Context, trustedExitRoot *etherman.GlobalExitRoot, dbTx pgx.Tx) (bool, error) {
-	addTrustedGerSQL := fmt.Sprintf(`
-		INSERT INTO sync.exit_root%[1]v (block_id, global_exit_root, exit_roots) 
+	const addTrustedGerSQL = `
+		INSERT INTO sync.exit_root (block_id, global_exit_root, exit_roots) 
 		VALUES (0, $1, $2)
-		ON CONFLICT ON CONSTRAINT UC DO NOTHING;`, p.tableSuffix)
+		ON CONFLICT ON CONSTRAINT UC DO NOTHING;`
 	res, err := p.getExecQuerier(dbTx).Exec(ctx, addTrustedGerSQL, trustedExitRoot.GlobalExitRoot, pq.Array([][]byte{trustedExitRoot.ExitRoots[0][:], trustedExitRoot.ExitRoots[1][:]}))
 	return res.RowsAffected() > 0, err
 }
@@ -212,10 +203,10 @@ func (p *PostgresStorage) GetClaim(ctx context.Context, depositCount, networkID 
 		claim  etherman.Claim
 		amount string
 	)
-	getClaimSQL := fmt.Sprintf(`
+	const getClaimSQL = `
 		SELECT index, orig_net, orig_addr, amount, dest_addr, block_id, c.network_id, tx_hash, b.received_at
-		FROM sync.claim%[1]v as c INNER JOIN sync.block%[1]v as b ON c.network_id = b.network_id AND c.block_id = b.id
-		WHERE index = $1 AND c.network_id = $2`, p.tableSuffix)
+		FROM sync.claim as c INNER JOIN sync.block as b ON c.network_id = b.network_id AND c.block_id = b.id
+		WHERE index = $1 AND c.network_id = $2`
 	err := p.getExecQuerier(dbTx).QueryRow(ctx, getClaimSQL, depositCount, networkID).Scan(&claim.Index, &claim.OriginalNetwork, &claim.OriginalAddress, &amount, &claim.DestinationAddress, &claim.BlockID, &claim.NetworkID, &claim.TxHash, &claim.Time)
 	if errors.Is(err, pgx.ErrNoRows) {
 		return nil, gerror.ErrStorageNotFound
@@ -230,13 +221,8 @@ func (p *PostgresStorage) GetDeposit(ctx context.Context, depositCounterUser uin
 		deposit etherman.Deposit
 		amount  string
 	)
-	getDepositSQL := fmt.Sprintf(`
-		SELECT leaf_type, orig_net, orig_addr, amount, dest_net, dest_addr, deposit_cnt, block_id, b.block_num, d.network_id, tx_hash, metadata, ready_for_claim
-		FROM sync.deposit%[1]v as d INNER JOIN sync.block%[1]v as b ON d.network_id = b.network_id AND d.block_id = b.id
-		WHERE d.network_id = $1 AND deposit_cnt = $2`, p.tableSuffix)
-	err := p.getExecQuerier(dbTx).QueryRow(ctx, getDepositSQL, networkID, depositCounterUser).Scan(
-		&deposit.LeafType, &deposit.OriginalNetwork, &deposit.OriginalAddress, &amount, &deposit.DestinationNetwork, &deposit.DestinationAddress,
-		&deposit.DepositCount, &deposit.BlockID, &deposit.BlockNumber, &deposit.NetworkID, &deposit.TxHash, &deposit.Metadata, &deposit.ReadyForClaim)
+	const getDepositSQL = "SELECT leaf_type, orig_net, orig_addr, amount, dest_net, dest_addr, deposit_cnt, block_id, b.block_num, d.network_id, tx_hash, metadata, ready_for_claim FROM sync.deposit as d INNER JOIN sync.block as b ON d.network_id = b.network_id AND d.block_id = b.id WHERE d.network_id = $1 AND deposit_cnt = $2"
+	err := p.getExecQuerier(dbTx).QueryRow(ctx, getDepositSQL, networkID, depositCounterUser).Scan(&deposit.LeafType, &deposit.OriginalNetwork, &deposit.OriginalAddress, &amount, &deposit.DestinationNetwork, &deposit.DestinationAddress, &deposit.DepositCount, &deposit.BlockID, &deposit.BlockNumber, &deposit.NetworkID, &deposit.TxHash, &deposit.Metadata, &deposit.ReadyForClaim)
 	if errors.Is(err, pgx.ErrNoRows) {
 		return nil, gerror.ErrStorageNotFound
 	}
@@ -260,7 +246,7 @@ func (p *PostgresStorage) GetLatestL1SyncedExitRoot(ctx context.Context, dbTx pg
 		ger       etherman.GlobalExitRoot
 		exitRoots [][]byte
 	)
-	getLatestL1SyncedExitRootSQL := fmt.Sprintf("SELECT block_id, global_exit_root, exit_roots FROM sync.exit_root%[1]v WHERE block_id > 0 ORDER BY id DESC LIMIT 1", p.tableSuffix)
+	const getLatestL1SyncedExitRootSQL = "SELECT block_id, global_exit_root, exit_roots FROM sync.exit_root WHERE block_id > 0 ORDER BY id DESC LIMIT 1"
 	err := p.getExecQuerier(dbTx).QueryRow(ctx, getLatestL1SyncedExitRootSQL).Scan(&ger.BlockID, &ger.GlobalExitRoot, pq.Array(&exitRoots))
 	if err != nil {
 		if errors.Is(err, pgx.ErrNoRows) {
@@ -278,7 +264,7 @@ func (p *PostgresStorage) GetLatestTrustedExitRoot(ctx context.Context, dbTx pgx
 		ger       etherman.GlobalExitRoot
 		exitRoots [][]byte
 	)
-	getLatestTrustedExitRootSQL := fmt.Sprintf("SELECT global_exit_root, exit_roots FROM sync.exit_root%[1]v WHERE block_id = 0 ORDER BY id DESC LIMIT 1", p.tableSuffix)
+	const getLatestTrustedExitRootSQL = "SELECT global_exit_root, exit_roots FROM sync.exit_root WHERE block_id = 0 ORDER BY id DESC LIMIT 1"
 	err := p.getExecQuerier(dbTx).QueryRow(ctx, getLatestTrustedExitRootSQL).Scan(&ger.GlobalExitRoot, pq.Array(&exitRoots))
 	if err != nil {
 		if errors.Is(err, pgx.ErrNoRows) {
@@ -292,11 +278,10 @@ func (p *PostgresStorage) GetLatestTrustedExitRoot(ctx context.Context, dbTx pgx
 
 // GetTokenWrapped gets a specific wrapped token.
 func (p *PostgresStorage) GetTokenWrapped(ctx context.Context, originalNetwork uint, originalTokenAddress common.Address, dbTx pgx.Tx) (*etherman.TokenWrapped, error) {
-	getWrappedTokenSQL := fmt.Sprintf("SELECT network_id, orig_net, orig_token_addr, wrapped_token_addr, block_id, name, symbol, decimals FROM sync.token_wrapped%[1]v WHERE orig_net = $1 AND orig_token_addr = $2", p.tableSuffix)
+	const getWrappedTokenSQL = "SELECT network_id, orig_net, orig_token_addr, wrapped_token_addr, block_id, name, symbol, decimals FROM sync.token_wrapped WHERE orig_net = $1 AND orig_token_addr = $2"
 
 	var token etherman.TokenWrapped
-	err := p.getExecQuerier(dbTx).QueryRow(ctx, getWrappedTokenSQL, originalNetwork, originalTokenAddress).Scan(&token.NetworkID, &token.OriginalNetwork,
-		&token.OriginalTokenAddress, &token.WrappedTokenAddress, &token.BlockID, &token.Name, &token.Symbol, &token.Decimals)
+	err := p.getExecQuerier(dbTx).QueryRow(ctx, getWrappedTokenSQL, originalNetwork, originalTokenAddress).Scan(&token.NetworkID, &token.OriginalNetwork, &token.OriginalTokenAddress, &token.WrappedTokenAddress, &token.BlockID, &token.Name, &token.Symbol, &token.Decimals)
 	if errors.Is(err, pgx.ErrNoRows) {
 		return nil, gerror.ErrStorageNotFound
 	}
@@ -315,7 +300,7 @@ func (p *PostgresStorage) GetTokenWrapped(ctx context.Context, originalNetwork u
 			if err != nil {
 				return nil, err
 			}
-			updateWrappedTokenSQL := fmt.Sprintf("UPDATE sync.token_wrapped%[1]v SET name = $3, symbol = $4, decimals = $5  WHERE orig_net = $1 AND orig_token_addr = $2", p.tableSuffix) //nolint: gosec
+			updateWrappedTokenSQL := "UPDATE sync.token_wrapped SET name = $3, symbol = $4, decimals = $5  WHERE orig_net = $1 AND orig_token_addr = $2" //nolint: gosec
 			_, err = p.getExecQuerier(dbTx).Exec(ctx, updateWrappedTokenSQL, originalNetwork, originalTokenAddress, tokenMetadata.Name, tokenMetadata.Symbol, tokenMetadata.Decimals)
 			if err != nil {
 				return nil, err
@@ -329,7 +314,7 @@ func (p *PostgresStorage) GetTokenWrapped(ctx context.Context, originalNetwork u
 // GetDepositCountByRoot gets the deposit count by the root.
 func (p *PostgresStorage) GetDepositCountByRoot(ctx context.Context, root []byte, network uint8, dbTx pgx.Tx) (uint, error) {
 	var depositCount uint
-	getDepositCountByRootSQL := fmt.Sprintf("SELECT deposit_cnt FROM mt.root%[1]v WHERE root = $1 AND network = $2", p.tableSuffix)
+	const getDepositCountByRootSQL = "SELECT deposit_cnt FROM mt.root WHERE root = $1 AND network = $2"
 	err := p.getExecQuerier(dbTx).QueryRow(ctx, getDepositCountByRootSQL, root, network).Scan(&depositCount)
 	if errors.Is(err, pgx.ErrNoRows) {
 		return 0, gerror.ErrStorageNotFound
@@ -340,7 +325,7 @@ func (p *PostgresStorage) GetDepositCountByRoot(ctx context.Context, root []byte
 // GetRoot gets root by the deposit count from the merkle tree.
 func (p *PostgresStorage) GetRoot(ctx context.Context, depositCnt uint, network uint, dbTx pgx.Tx) ([]byte, error) {
 	var root []byte
-	getRootByDepositCntSQL := fmt.Sprintf("SELECT root FROM mt.root%[1]v WHERE deposit_cnt = $1 AND network = $2", p.tableSuffix)
+	const getRootByDepositCntSQL = "SELECT root FROM mt.root WHERE deposit_cnt = $1 AND network = $2"
 	err := p.getExecQuerier(dbTx).QueryRow(ctx, getRootByDepositCntSQL, depositCnt, network).Scan(&root)
 	if errors.Is(err, pgx.ErrNoRows) {
 		return nil, gerror.ErrStorageNotFound
@@ -350,14 +335,14 @@ func (p *PostgresStorage) GetRoot(ctx context.Context, depositCnt uint, network 
 
 // SetRoot store the root with deposit count to the storage.
 func (p *PostgresStorage) SetRoot(ctx context.Context, root []byte, depositID uint64, depositCnt uint, network uint, dbTx pgx.Tx) error {
-	setRootSQL := fmt.Sprintf("INSERT INTO mt.root%[1]v (root, deposit_id, deposit_cnt, network) VALUES ($1, $2, $3, $4);", p.tableSuffix)
+	const setRootSQL = "INSERT INTO mt.root (root, deposit_id, deposit_cnt, network) VALUES ($1, $2, $3, $4);"
 	_, err := p.getExecQuerier(dbTx).Exec(ctx, setRootSQL, root, depositID, depositCnt, network)
 	return err
 }
 
 // Get gets value of key from the merkle tree.
 func (p *PostgresStorage) Get(ctx context.Context, key []byte, dbTx pgx.Tx) ([][]byte, error) {
-	getValueByKeySQL := fmt.Sprintf("SELECT value FROM mt.rht%[1]v WHERE key = $1", p.tableSuffix)
+	const getValueByKeySQL = "SELECT value FROM mt.rht WHERE key = $1"
 	var data [][]byte
 	err := p.getExecQuerier(dbTx).QueryRow(ctx, getValueByKeySQL, key).Scan(pq.Array(&data))
 	if errors.Is(err, pgx.ErrNoRows) {
@@ -370,21 +355,21 @@ func (p *PostgresStorage) Get(ctx context.Context, key []byte, dbTx pgx.Tx) ([][
 // If record with such a key already exists its assumed that the value is correct,
 // because it's a reverse hash table, and the key is a hash of the value
 func (p *PostgresStorage) Set(ctx context.Context, key []byte, value [][]byte, depositID uint64, dbTx pgx.Tx) error {
-	setNodeSQL := fmt.Sprintf("INSERT INTO mt.rht%[1]v (deposit_id, key, value) VALUES ($1, $2, $3)", p.tableSuffix)
+	const setNodeSQL = "INSERT INTO mt.rht (deposit_id, key, value) VALUES ($1, $2, $3)"
 	_, err := p.getExecQuerier(dbTx).Exec(ctx, setNodeSQL, depositID, key, pq.Array(value))
 	return err
 }
 
 // BulkSet is similar to Set, but it inserts multiple key-value pairs into the db.
 func (p *PostgresStorage) BulkSet(ctx context.Context, rows [][]interface{}, dbTx pgx.Tx) error {
-	_, err := p.getExecQuerier(dbTx).CopyFrom(ctx, pgx.Identifier{"mt", "rht" + p.tableSuffix.Get()}, []string{"key", "value", "deposit_id"}, pgx.CopyFromRows(rows))
+	_, err := p.getExecQuerier(dbTx).CopyFrom(ctx, pgx.Identifier{"mt", "rht"}, []string{"key", "value", "deposit_id"}, pgx.CopyFromRows(rows))
 	return err
 }
 
 // GetLastDepositCount gets the last deposit count from the merkle tree.
 func (p *PostgresStorage) GetLastDepositCount(ctx context.Context, network uint, dbTx pgx.Tx) (uint, error) {
 	var depositCnt int64
-	getLastDepositCountSQL := fmt.Sprintf("SELECT coalesce(MAX(deposit_cnt), -1) FROM mt.root%[1]v WHERE network = $1", p.tableSuffix)
+	const getLastDepositCountSQL = "SELECT coalesce(MAX(deposit_cnt), -1) FROM mt.root WHERE network = $1"
 	err := p.getExecQuerier(dbTx).QueryRow(ctx, getLastDepositCountSQL, network).Scan(&depositCnt)
 	if err != nil {
 		return 0, nil
@@ -397,7 +382,7 @@ func (p *PostgresStorage) GetLastDepositCount(ctx context.Context, network uint,
 
 // GetClaimCount gets the claim count for the destination address.
 func (p *PostgresStorage) GetClaimCount(ctx context.Context, destAddr string, dbTx pgx.Tx) (uint64, error) {
-	getClaimCountSQL := fmt.Sprintf("SELECT COUNT(*) FROM sync.claim%[1]v WHERE dest_addr = $1", p.tableSuffix)
+	const getClaimCountSQL = "SELECT COUNT(*) FROM sync.claim WHERE dest_addr = $1"
 	var claimCount uint64
 	err := p.getExecQuerier(dbTx).QueryRow(ctx, getClaimCountSQL, common.FromHex(destAddr)).Scan(&claimCount)
 	if errors.Is(err, pgx.ErrNoRows) {
@@ -408,7 +393,7 @@ func (p *PostgresStorage) GetClaimCount(ctx context.Context, destAddr string, db
 
 // GetClaims gets the claim list which be smaller than index.
 func (p *PostgresStorage) GetClaims(ctx context.Context, destAddr string, limit uint, offset uint, dbTx pgx.Tx) ([]*etherman.Claim, error) {
-	getClaimsSQL := fmt.Sprintf("SELECT index, orig_net, orig_addr, amount, dest_addr, block_id, network_id, tx_hash FROM sync.claim%[1]v WHERE dest_addr = $1 ORDER BY block_id DESC LIMIT $2 OFFSET $3", p.tableSuffix)
+	const getClaimsSQL = "SELECT index, orig_net, orig_addr, amount, dest_addr, block_id, network_id, tx_hash FROM sync.claim WHERE dest_addr = $1 ORDER BY block_id DESC LIMIT $2 OFFSET $3"
 	rows, err := p.getExecQuerier(dbTx).Query(ctx, getClaimsSQL, common.FromHex(destAddr), limit, offset)
 	if err != nil {
 		return nil, err
@@ -432,160 +417,17 @@ func (p *PostgresStorage) GetClaims(ctx context.Context, destAddr string, limit 
 
 // GetDeposits gets the deposit list which be smaller than depositCount.
 func (p *PostgresStorage) GetDeposits(ctx context.Context, destAddr string, limit uint, offset uint, dbTx pgx.Tx) ([]*etherman.Deposit, error) {
-	getDepositsSQL := fmt.Sprintf(`
+	const getDepositsSQL = `
 		SELECT d.id, leaf_type, orig_net, orig_addr, amount, dest_net, dest_addr, deposit_cnt, block_id, b.block_num, d.network_id, tx_hash, metadata, ready_for_claim, b.received_at
-		FROM sync.deposit%[1]v as d INNER JOIN sync.block%[1]v as b ON d.network_id = b.network_id AND d.block_id = b.id
+		FROM sync.deposit as d INNER JOIN sync.block as b ON d.network_id = b.network_id AND d.block_id = b.id
 		WHERE dest_addr = $1
-		ORDER BY d.block_id DESC, d.deposit_cnt DESC LIMIT $2 OFFSET $3`, p.tableSuffix)
-	rows, err := p.getExecQuerier(dbTx).Query(ctx, getDepositsSQL, common.FromHex(destAddr), limit, offset)
-	return p.convertDepositBase(rows, err)
-}
-
-// GetDepositsWithLeafType gets the deposit list which be smaller than depositCount.
-func (p *PostgresStorage) GetDepositsWithLeafType(ctx context.Context, destAddr string, limit uint, offset uint, leafType uint, dbTx pgx.Tx) ([]*etherman.Deposit, error) {
-	getDepositsSQL := fmt.Sprintf(`
-		SELECT d.id, leaf_type, orig_net, orig_addr, amount, dest_net, dest_addr, deposit_cnt, block_id, b.block_num, d.network_id, tx_hash, metadata, ready_for_claim, b.received_at
-		FROM sync.deposit%[1]v as d INNER JOIN sync.block%[1]v as b ON d.network_id = b.network_id AND d.block_id = b.id
-		WHERE dest_addr = $1 AND leaf_type = $4
-		ORDER BY d.block_id DESC, d.deposit_cnt DESC LIMIT $2 OFFSET $3`, p.tableSuffix)
-	rows, err := p.getExecQuerier(dbTx).Query(ctx, getDepositsSQL, common.FromHex(destAddr), limit, offset, leafType)
-	return p.convertDepositBase(rows, err)
-}
-
-func (p *PostgresStorage) convertDepositBase(rows pgx.Rows, err error) ([]*etherman.Deposit, error) {
-	if err != nil {
-		return nil, err
-	}
-
-	deposits := make([]*etherman.Deposit, 0, len(rows.RawValues()))
-
-	for rows.Next() {
-		var (
-			deposit etherman.Deposit
-			amount  string
-		)
-		err = rows.Scan(&deposit.Id, &deposit.LeafType, &deposit.OriginalNetwork, &deposit.OriginalAddress, &amount, &deposit.DestinationNetwork, &deposit.DestinationAddress,
-			&deposit.DepositCount, &deposit.BlockID, &deposit.BlockNumber, &deposit.NetworkID, &deposit.TxHash, &deposit.Metadata, &deposit.ReadyForClaim, &deposit.Time)
-		if err != nil {
-			return nil, err
-		}
-		deposit.Amount, _ = new(big.Int).SetString(amount, 10) //nolint:gomnd
-		deposits = append(deposits, &deposit)
-	}
-
-	return deposits, nil
-}
-
-// GetDepositByHash returns a deposit from a specific account and tx hash
-func (p *PostgresStorage) GetDepositByHash(ctx context.Context, destAddr string, networkID uint, txHash string, dbTx pgx.Tx) (*etherman.Deposit, error) {
-	var (
-		deposit etherman.Deposit
-		amount  string
-	)
-	getDepositSQL := fmt.Sprintf(`
-		SELECT leaf_type, orig_net, orig_addr, amount, dest_net, dest_addr, deposit_cnt, block_id, b.block_num, d.network_id, tx_hash, metadata, ready_for_claim
-		FROM sync.deposit%[1]v as d INNER JOIN sync.block%[1]v as b ON d.network_id = b.network_id AND d.block_id = b.id
-		WHERE d.dest_addr = $1 AND d.network_id = $2 AND d.tx_hash = $3`, p.tableSuffix)
-	err := p.getExecQuerier(dbTx).QueryRow(ctx, getDepositSQL, common.HexToAddress(destAddr), networkID, common.HexToHash(txHash)).Scan(
-		&deposit.LeafType, &deposit.OriginalNetwork, &deposit.OriginalAddress, &amount, &deposit.DestinationNetwork, &deposit.DestinationAddress,
-		&deposit.DepositCount, &deposit.BlockID, &deposit.BlockNumber, &deposit.NetworkID, &deposit.TxHash, &deposit.Metadata, &deposit.ReadyForClaim)
-	if errors.Is(err, pgx.ErrNoRows) {
-		return nil, gerror.ErrStorageNotFound
-	}
-	deposit.Amount, _ = new(big.Int).SetString(amount, 10) //nolint:gomnd
-
-	return &deposit, err
-}
-
-// GetPendingTransactions gets all the deposit transactions of a user that have not been claimed
-func (p *PostgresStorage) GetPendingTransactions(ctx context.Context, destAddr string, limit uint, offset uint, leafType uint, dbTx pgx.Tx) ([]*etherman.Deposit, error) {
-	getDepositsSQL := fmt.Sprintf(`SELECT d.id, leaf_type, orig_net, orig_addr, amount, dest_net, dest_addr, deposit_cnt, block_id, b.block_num, d.network_id, tx_hash, metadata, ready_for_claim, b.received_at
-		FROM sync.deposit%[1]v as d INNER JOIN sync.block%[1]v as b ON d.network_id = b.network_id AND d.block_id = b.id
-		WHERE dest_addr = $1 AND leaf_type = $4 AND NOT EXISTS
-			(SELECT 1 FROM sync.claim%[1]v as c WHERE c.index = d.deposit_cnt AND c.network_id = d.dest_net)
-		ORDER BY d.block_id DESC, d.deposit_cnt DESC LIMIT $2 OFFSET $3`, p.tableSuffix)
-
-	return p.getDepositList(ctx, getDepositsSQL, dbTx, common.FromHex(destAddr), limit, offset, leafType)
-}
-
-// GetNotReadyTransactions returns all the deposit transactions with ready_for_claim = false
-func (p *PostgresStorage) GetNotReadyTransactions(ctx context.Context, limit uint, offset uint, dbTx pgx.Tx) ([]*etherman.Deposit, error) {
-	getDepositsSQL := fmt.Sprintf(`SELECT d.id, leaf_type, orig_net, orig_addr, amount, dest_net, dest_addr, deposit_cnt, block_id, b.block_num, d.network_id, tx_hash, metadata, ready_for_claim, b.received_at
-		FROM sync.deposit%[1]v as d INNER JOIN sync.block%[1]v as b ON d.network_id = b.network_id AND d.block_id = b.id
-		WHERE ready_for_claim = false
-		ORDER BY d.block_id DESC, d.deposit_cnt DESC LIMIT $1 OFFSET $2`, p.tableSuffix)
-
-	return p.getDepositList(ctx, getDepositsSQL, dbTx, limit, offset)
-}
-
-func (p *PostgresStorage) GetReadyPendingTransactions(ctx context.Context, networkID uint, leafType uint, limit uint, offset uint, minReadyTime time.Time, dbTx pgx.Tx) ([]*etherman.Deposit, error) {
-	getDepositsSQL := fmt.Sprintf(`SELECT d.id, leaf_type, orig_net, orig_addr, amount, dest_net, dest_addr, deposit_cnt, block_id, b.block_num, d.network_id, tx_hash, metadata, ready_for_claim, b.received_at
-		FROM sync.deposit%[1]v as d INNER JOIN sync.block%[1]v as b ON d.network_id = b.network_id AND d.block_id = b.id
-		WHERE d.network_id = $1 AND ready_for_claim = true AND leaf_type = $4 AND ready_time >= $5 AND NOT EXISTS
-			(SELECT 1 FROM sync.claim%[1]v as c WHERE c.index = d.deposit_cnt AND c.network_id = d.dest_net)
-		ORDER BY d.block_id DESC, d.deposit_cnt DESC LIMIT $2 OFFSET $3`, p.tableSuffix)
-
-	return p.getDepositList(ctx, getDepositsSQL, dbTx, networkID, limit, offset, leafType, minReadyTime)
-}
-
-func (p *PostgresStorage) getDepositList(ctx context.Context, sql string, dbTx pgx.Tx, args ...interface{}) ([]*etherman.Deposit, error) {
-	rows, err := p.getExecQuerier(dbTx).Query(ctx, sql, args...)
-	if err != nil {
-		return nil, err
-	}
-
-	deposits := make([]*etherman.Deposit, 0, len(rows.RawValues()))
-
-	for rows.Next() {
-		var (
-			deposit etherman.Deposit
-			amount  string
-		)
-		err = rows.Scan(&deposit.Id, &deposit.LeafType, &deposit.OriginalNetwork, &deposit.OriginalAddress, &amount, &deposit.DestinationNetwork, &deposit.DestinationAddress,
-			&deposit.DepositCount, &deposit.BlockID, &deposit.BlockNumber, &deposit.NetworkID, &deposit.TxHash, &deposit.Metadata, &deposit.ReadyForClaim, &deposit.Time)
-		if err != nil {
-			return nil, err
-		}
-		deposit.Amount, _ = new(big.Int).SetString(amount, 10) //nolint:gomnd
-		deposits = append(deposits, &deposit)
-	}
-
-	return deposits, nil
-}
-
-func (p *PostgresStorage) GetNotReadyTransactionsWithBlockRange(ctx context.Context, networkID uint, minBlockNum, maxBlockNum uint64, limit, offset uint, dbTx pgx.Tx) ([]*etherman.Deposit, error) {
-	getDepositsSQL := fmt.Sprintf(`SELECT d.id, leaf_type, orig_net, orig_addr, amount, dest_net, dest_addr, deposit_cnt, block_id, b.block_num, d.network_id, tx_hash, metadata, ready_for_claim, b.received_at
-		FROM sync.deposit%[1]v as d INNER JOIN sync.block%[1]v as b ON d.network_id = b.network_id AND d.block_id = b.id
-		WHERE b.network_id = $1 AND ready_for_claim = false AND b.block_num >= $2 AND b.block_num <= $3
-		ORDER BY d.block_id DESC, d.deposit_cnt DESC LIMIT $4 OFFSET $5`, p.tableSuffix)
-
-	rows, err := p.getExecQuerier(dbTx).Query(ctx, getDepositsSQL, networkID, minBlockNum, maxBlockNum, limit, offset)
-	if err != nil {
-		return nil, err
-	}
-
-	deposits := make([]*etherman.Deposit, 0, len(rows.RawValues()))
-
-	for rows.Next() {
-		var (
-			deposit etherman.Deposit
-			amount  string
-		)
-		err = rows.Scan(&deposit.Id, &deposit.LeafType, &deposit.OriginalNetwork, &deposit.OriginalAddress, &amount, &deposit.DestinationNetwork, &deposit.DestinationAddress,
-			&deposit.DepositCount, &deposit.BlockID, &deposit.BlockNumber, &deposit.NetworkID, &deposit.TxHash, &deposit.Metadata, &deposit.ReadyForClaim, &deposit.Time)
-		if err != nil {
-			return nil, err
-		}
-		deposit.Amount, _ = new(big.Int).SetString(amount, 10) //nolint:gomnd
-		deposits = append(deposits, &deposit)
-	}
-
-	return deposits, nil
+		ORDER BY d.block_id DESC, d.deposit_cnt DESC LIMIT $2 OFFSET $3`
+	return p.getDepositList(ctx, getDepositsSQL, dbTx, common.FromHex(destAddr), limit, offset)
 }
 
 // GetDepositCount gets the deposit count for the destination address.
 func (p *PostgresStorage) GetDepositCount(ctx context.Context, destAddr string, dbTx pgx.Tx) (uint64, error) {
-	getDepositCountSQL := fmt.Sprintf("SELECT COUNT(*) FROM sync.deposit%[1]v WHERE dest_addr = $1", p.tableSuffix)
+	const getDepositCountSQL = "SELECT COUNT(*) FROM sync.deposit WHERE dest_addr = $1"
 	var depositCount uint64
 	err := p.getExecQuerier(dbTx).QueryRow(ctx, getDepositCountSQL, common.FromHex(destAddr)).Scan(&depositCount)
 	return depositCount, err
@@ -593,17 +435,18 @@ func (p *PostgresStorage) GetDepositCount(ctx context.Context, destAddr string, 
 
 // UpdateBlocksForTesting updates the hash of blocks.
 func (p *PostgresStorage) UpdateBlocksForTesting(ctx context.Context, networkID uint, blockNum uint64, dbTx pgx.Tx) error {
-	updateBlocksSQL := fmt.Sprintf("UPDATE sync.block%[1]v SET block_hash = SUBSTRING(block_hash FROM 1 FOR LENGTH(block_hash)-1) || '\x61' WHERE network_id = $1 AND block_num >= $2", p.tableSuffix)
+	const updateBlocksSQL = "UPDATE sync.block SET block_hash = SUBSTRING(block_hash FROM 1 FOR LENGTH(block_hash)-1) || '\x61' WHERE network_id = $1 AND block_num >= $2"
 	_, err := p.getExecQuerier(dbTx).Exec(ctx, updateBlocksSQL, networkID, blockNum)
 	return err
 }
 
-// GetL1Deposits get the L1 deposits remain to be ready_for_claim
-func (p *PostgresStorage) GetL1Deposits(ctx context.Context, exitRoot []byte, dbTx pgx.Tx) ([]*etherman.Deposit, error) {
-	updateDepositsStatusSQL := fmt.Sprintf(`Select d.id, leaf_type, orig_net, orig_addr, amount, dest_net, dest_addr, deposit_cnt, block_id, b.block_num, d.network_id, tx_hash, metadata, ready_for_claim, b.received_at
-			FROM sync.deposit%[1]v as d INNER JOIN sync.block%[1]v as b ON d.network_id = b.network_id AND d.block_id = b.id
-			WHERE deposit_cnt <= (SELECT deposit_cnt FROM mt.root%[1]v WHERE root = $1 AND network = 0) 
-			AND d.network_id = 0 AND ready_for_claim = false`, p.tableSuffix)
+// UpdateL1DepositsStatus updates the ready_for_claim status of L1 deposits.
+func (p *PostgresStorage) UpdateL1DepositsStatus(ctx context.Context, exitRoot []byte, dbTx pgx.Tx) ([]*etherman.Deposit, error) {
+	const updateDepositsStatusSQL = `UPDATE sync.deposit SET ready_for_claim = true 
+		WHERE deposit_cnt <=
+			(SELECT deposit_cnt FROM mt.root WHERE root = $1 AND network = 0) 
+			AND network_id = 0 AND ready_for_claim = false
+			RETURNING leaf_type, orig_net, orig_addr, amount, dest_net, dest_addr, deposit_cnt, block_id, network_id, tx_hash, metadata, ready_for_claim;`
 	rows, err := p.getExecQuerier(dbTx).Query(ctx, updateDepositsStatusSQL, exitRoot)
 	if err != nil {
 		return nil, err
@@ -615,74 +458,7 @@ func (p *PostgresStorage) GetL1Deposits(ctx context.Context, exitRoot []byte, db
 			deposit etherman.Deposit
 			amount  string
 		)
-		err = rows.Scan(&deposit.Id, &deposit.LeafType, &deposit.OriginalNetwork, &deposit.OriginalAddress, &amount, &deposit.DestinationNetwork, &deposit.DestinationAddress,
-			&deposit.DepositCount, &deposit.BlockID, &deposit.BlockNumber, &deposit.NetworkID, &deposit.TxHash, &deposit.Metadata, &deposit.ReadyForClaim, &deposit.Time)
-		if err != nil {
-			return nil, err
-		}
-		deposit.Amount, _ = new(big.Int).SetString(amount, 10) //nolint:gomnd
-		deposits = append(deposits, &deposit)
-	}
-	return deposits, nil
-}
-
-// UpdateL1DepositsStatus updates the ready_for_claim status of L1 deposits.
-func (p *PostgresStorage) UpdateL1DepositsStatus(ctx context.Context, exitRoot []byte, eventTime time.Time, dbTx pgx.Tx) ([]*etherman.Deposit, error) {
-	updateDepositsStatusSQL := fmt.Sprintf(`WITH d AS (UPDATE sync.deposit%[1]v SET ready_for_claim = true, ready_time = $1 
-		WHERE deposit_cnt <=
-			(SELECT deposit_cnt FROM mt.root%[1]v WHERE root = $2 AND network = 0) 
-			AND network_id = 0 AND ready_for_claim = false
-			RETURNING *)
-		SELECT d.id, leaf_type, orig_net, orig_addr, amount, dest_net, dest_addr, deposit_cnt, block_id, b.block_num, d.network_id, tx_hash, metadata, ready_for_claim, b.received_at
-		FROM d INNER JOIN sync.block%[1]v as b ON d.network_id = b.network_id AND d.block_id = b.id`, p.tableSuffix)
-	rows, err := p.getExecQuerier(dbTx).Query(ctx, updateDepositsStatusSQL, eventTime, exitRoot)
-	if err != nil {
-		return nil, err
-	}
-
-	deposits := make([]*etherman.Deposit, 0, len(rows.RawValues()))
-	for rows.Next() {
-		var (
-			deposit etherman.Deposit
-			amount  string
-		)
-		err = rows.Scan(&deposit.Id, &deposit.LeafType, &deposit.OriginalNetwork, &deposit.OriginalAddress, &amount, &deposit.DestinationNetwork, &deposit.DestinationAddress,
-			&deposit.DepositCount, &deposit.BlockID, &deposit.BlockNumber, &deposit.NetworkID, &deposit.TxHash, &deposit.Metadata, &deposit.ReadyForClaim, &deposit.Time)
-		if err != nil {
-			return nil, err
-		}
-		deposit.Amount, _ = new(big.Int).SetString(amount, 10) //nolint:gomnd
-		deposits = append(deposits, &deposit)
-	}
-	return deposits, nil
-}
-
-func (p *PostgresStorage) UpdateL1DepositStatus(ctx context.Context, depositCount uint, eventTime time.Time, dbTx pgx.Tx) error {
-	updateDepositStatusSQL := fmt.Sprintf(`UPDATE sync.deposit%[1]v SET ready_for_claim = true, ready_time = $1 
-		WHERE deposit_cnt = $2 And network_id = 0`, p.tableSuffix)
-	_, err := p.getExecQuerier(dbTx).Exec(ctx, updateDepositStatusSQL, eventTime, depositCount)
-	return err
-}
-
-// UpdateL2DepositsStatus updates the ready_for_claim status of L2 deposits. and return deposit list
-func (p *PostgresStorage) UpdateL2DepositsStatusWithBackDeposits(ctx context.Context, exitRoot []byte, eventTime time.Time, dbTx pgx.Tx) ([]*etherman.Deposit, error) {
-	updateDepositsStatusSQL := fmt.Sprintf(`UPDATE sync.deposit%[1]v SET ready_for_claim = true, ready_time = $1
-		WHERE deposit_cnt <=
-			(SELECT deposit_cnt FROM mt.root%[1]v WHERE root = $2 AND network = 1)
-			AND network_id = 1 AND ready_for_claim = false
-			RETURNING leaf_type, orig_net, orig_addr, amount, dest_net, dest_addr, deposit_cnt, block_id, network_id, tx_hash, metadata, ready_for_claim;`, p.tableSuffix)
-	rows, err := p.getExecQuerier(dbTx).Query(ctx, updateDepositsStatusSQL, eventTime, exitRoot)
-	if err != nil {
-		return nil, err
-	}
-	deposits := make([]*etherman.Deposit, 0, len(rows.RawValues()))
-	for rows.Next() {
-		var (
-			deposit etherman.Deposit
-			amount  string
-		)
-		err = rows.Scan(&deposit.LeafType, &deposit.OriginalNetwork, &deposit.OriginalAddress, &amount, &deposit.DestinationNetwork, &deposit.DestinationAddress,
-			&deposit.DepositCount, &deposit.BlockID, &deposit.NetworkID, &deposit.TxHash, &deposit.Metadata, &deposit.ReadyForClaim)
+		err = rows.Scan(&deposit.LeafType, &deposit.OriginalNetwork, &deposit.OriginalAddress, &amount, &deposit.DestinationNetwork, &deposit.DestinationAddress, &deposit.DepositCount, &deposit.BlockID, &deposit.NetworkID, &deposit.TxHash, &deposit.Metadata, &deposit.ReadyForClaim)
 		if err != nil {
 			return nil, err
 		}
@@ -693,27 +469,27 @@ func (p *PostgresStorage) UpdateL2DepositsStatusWithBackDeposits(ctx context.Con
 }
 
 // UpdateL2DepositsStatus updates the ready_for_claim status of L2 deposits.
-func (p *PostgresStorage) UpdateL2DepositsStatus(ctx context.Context, exitRoot []byte, eventTime time.Time, networkID uint, dbTx pgx.Tx) error {
-	updateDepositsStatusSQL := fmt.Sprintf(`UPDATE sync.deposit%[1]v SET ready_for_claim = true, ready_time = $1
+func (p *PostgresStorage) UpdateL2DepositsStatus(ctx context.Context, exitRoot []byte, networkID uint, dbTx pgx.Tx) error {
+	const updateDepositsStatusSQL = `UPDATE sync.deposit SET ready_for_claim = true
 		WHERE deposit_cnt <=
-			(SELECT deposit_cnt FROM mt.root%[1]v WHERE root = $2 AND network = $3)
-			AND network_id = $3 AND ready_for_claim = false;`, p.tableSuffix)
-	_, err := p.getExecQuerier(dbTx).Exec(ctx, updateDepositsStatusSQL, eventTime, exitRoot, networkID)
+			(SELECT deposit_cnt FROM mt.root WHERE root = $1 AND network = $2)
+			AND network_id = $2 AND ready_for_claim = false;`
+	_, err := p.getExecQuerier(dbTx).Exec(ctx, updateDepositsStatusSQL, exitRoot, networkID)
 	return err
 }
 
 // AddClaimTx adds a claim monitored transaction to the storage.
 func (p *PostgresStorage) AddClaimTx(ctx context.Context, mTx ctmtypes.MonitoredTx, dbTx pgx.Tx) error {
-	addMonitoredTxSQL := fmt.Sprintf(`INSERT INTO sync.monitored_txs%[1]v 
+	const addMonitoredTxSQL = `INSERT INTO sync.monitored_txs 
 		(deposit_id, from_addr, to_addr, nonce, value, data, gas, status, history, created_at, updated_at)
-		VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11)`, p.tableSuffix)
+		VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11)`
 	_, err := p.getExecQuerier(dbTx).Exec(ctx, addMonitoredTxSQL, mTx.DepositID, mTx.From, mTx.To, mTx.Nonce, mTx.Value.String(), mTx.Data, mTx.Gas, mTx.Status, pq.Array(mTx.HistoryHashSlice()), time.Now().UTC(), time.Now().UTC())
 	return err
 }
 
 // UpdateClaimTx updates a claim monitored transaction in the storage.
 func (p *PostgresStorage) UpdateClaimTx(ctx context.Context, mTx ctmtypes.MonitoredTx, dbTx pgx.Tx) error {
-	updateMonitoredTxSQL := fmt.Sprintf(`UPDATE sync.monitored_txs%[1]v 
+	const updateMonitoredTxSQL = `UPDATE sync.monitored_txs 
 		SET from_addr = $2
 		, to_addr = $3
 		, nonce = $4
@@ -723,14 +499,14 @@ func (p *PostgresStorage) UpdateClaimTx(ctx context.Context, mTx ctmtypes.Monito
 		, status = $8
 		, history = $9
 		, updated_at = $10
-		WHERE deposit_id = $1`, p.tableSuffix)
+		WHERE deposit_id = $1`
 	_, err := p.getExecQuerier(dbTx).Exec(ctx, updateMonitoredTxSQL, mTx.DepositID, mTx.From, mTx.To, mTx.Nonce, mTx.Value.String(), mTx.Data, mTx.Gas, mTx.Status, pq.Array(mTx.HistoryHashSlice()), time.Now().UTC())
 	return err
 }
 
 // GetClaimTxsByStatus gets the monitored transactions by status.
 func (p *PostgresStorage) GetClaimTxsByStatus(ctx context.Context, statuses []ctmtypes.MonitoredTxStatus, dbTx pgx.Tx) ([]ctmtypes.MonitoredTx, error) {
-	getMonitoredTxsSQL := fmt.Sprintf("SELECT deposit_id, from_addr, to_addr, nonce, value, data, gas, status, history, created_at, updated_at FROM sync.monitored_txs%[1]v WHERE status = ANY($1) ORDER BY created_at ASC", p.tableSuffix)
+	const getMonitoredTxsSQL = "SELECT deposit_id, from_addr, to_addr, nonce, value, data, gas, status, history, created_at, updated_at FROM sync.monitored_txs WHERE status = ANY($1) ORDER BY created_at ASC"
 	rows, err := p.getExecQuerier(dbTx).Query(ctx, getMonitoredTxsSQL, pq.Array(statuses))
 	if errors.Is(err, pgx.ErrNoRows) {
 		return []ctmtypes.MonitoredTx{}, nil
@@ -760,122 +536,9 @@ func (p *PostgresStorage) GetClaimTxsByStatus(ctx context.Context, statuses []ct
 	return mTxs, nil
 }
 
-func (p *PostgresStorage) GetClaimTxsByStatusWithLimit(ctx context.Context, statuses []ctmtypes.MonitoredTxStatus, limit uint, offset uint, dbTx pgx.Tx) ([]ctmtypes.MonitoredTx, error) {
-	getMonitoredTxsSQL := fmt.Sprintf("SELECT deposit_id, from_addr, to_addr, nonce, value, data, gas, status, history, created_at, updated_at FROM sync.monitored_txs%[1]v WHERE status = ANY($1) ORDER BY created_at DESC LIMIT $2 OFFSET $3", p.tableSuffix)
-	rows, err := p.getExecQuerier(dbTx).Query(ctx, getMonitoredTxsSQL, pq.Array(statuses), limit, offset)
-	if errors.Is(err, pgx.ErrNoRows) {
-		return []ctmtypes.MonitoredTx{}, nil
-	} else if err != nil {
-		return nil, err
-	}
-
-	mTxs := make([]ctmtypes.MonitoredTx, 0, len(rows.RawValues()))
-	for rows.Next() {
-		var (
-			value   string
-			history [][]byte
-		)
-		mTx := ctmtypes.MonitoredTx{}
-		err = rows.Scan(&mTx.DepositID, &mTx.From, &mTx.To, &mTx.Nonce, &value, &mTx.Data, &mTx.Gas, &mTx.Status, pq.Array(&history), &mTx.CreatedAt, &mTx.UpdatedAt)
-		if err != nil {
-			return mTxs, err
-		}
-		mTx.Value, _ = new(big.Int).SetString(value, 10) //nolint:gomnd
-		mTx.History = make(map[common.Hash]bool)
-		for _, h := range history {
-			mTx.History[common.BytesToHash(h)] = true
-		}
-		mTxs = append(mTxs, mTx)
-	}
-
-	return mTxs, nil
-}
-
-// GetClaimTxById gets the monitored transactions by id (depositCount)
-func (p *PostgresStorage) GetClaimTxById(ctx context.Context, id uint, dbTx pgx.Tx) (*ctmtypes.MonitoredTx, error) {
-	getClaimSql := fmt.Sprintf("SELECT deposit_id, from_addr, to_addr, nonce, value, data, gas, status, history, created_at, updated_at FROM sync.monitored_txs%[1]v WHERE id = $1", p.tableSuffix)
-	var (
-		value   string
-		history [][]byte
-		mTx     = &ctmtypes.MonitoredTx{}
-	)
-	err := p.getExecQuerier(dbTx).QueryRow(ctx, getClaimSql, id).
-		Scan(&mTx.DepositID, &mTx.From, &mTx.To, &mTx.Nonce, &value, &mTx.Data, &mTx.Gas, &mTx.Status, pq.Array(&history), &mTx.CreatedAt, &mTx.UpdatedAt)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return nil, gerror.ErrStorageNotFound
-		}
-		return nil, err
-	}
-	mTx.Value, _ = new(big.Int).SetString(value, 10) //nolint:gomnd
-	mTx.History = make(map[common.Hash]bool)
-	for _, h := range history {
-		mTx.History[common.BytesToHash(h)] = true
-	}
-
-	return mTx, nil
-}
-
 // UpdateDepositsStatusForTesting updates the ready_for_claim status of all deposits for testing.
 func (p *PostgresStorage) UpdateDepositsStatusForTesting(ctx context.Context, dbTx pgx.Tx) error {
-	updateDepositsStatusSQL := fmt.Sprintf("UPDATE sync.deposit%[1]v SET ready_for_claim = true;", p.tableSuffix)
+	const updateDepositsStatusSQL = "UPDATE sync.deposit SET ready_for_claim = true;"
 	_, err := p.getExecQuerier(dbTx).Exec(ctx, updateDepositsStatusSQL)
 	return err
-}
-
-// GetAllMainCoins returns all the coin info from the main_coins table
-func (p *PostgresStorage) GetAllMainCoins(ctx context.Context, limit uint, offset uint, dbTx pgx.Tx) ([]*pb.CoinInfo, error) {
-	getCoinsSQL := fmt.Sprintf(`SELECT symbol, name, decimals, encode(address, 'hex'), chain_id, network_id, logo_link
-		FROM common.main_coins%[1]v WHERE is_deleted = false ORDER BY id LIMIT $1 OFFSET $2`, p.tableSuffix)
-	rows, err := p.getExecQuerier(dbTx).Query(ctx, getCoinsSQL, limit, offset)
-	if err != nil {
-		return nil, err
-	}
-
-	var result []*pb.CoinInfo
-	for rows.Next() {
-		coin := &pb.CoinInfo{}
-		err = rows.Scan(&coin.Symbol, &coin.Name, &coin.Decimals, &coin.Address, &coin.ChainId, &coin.NetworkId, &coin.LogoLink)
-		if err != nil {
-			log.Errorf("GetAllMainCoins scan row error[%v]", err)
-			return nil, err
-		}
-		if coin.Address != "" {
-			coin.Address = "0x" + coin.Address
-		}
-		result = append(result, coin)
-	}
-	return result, nil
-}
-
-// GetLatestReadyDeposits returns the latest deposit transactions with ready_for_claim = true
-func (p *PostgresStorage) GetLatestReadyDeposits(ctx context.Context, networkID uint, limit uint, dbTx pgx.Tx) ([]*etherman.Deposit, error) {
-	getDepositsSQL := fmt.Sprintf(`
-		SELECT d.id, leaf_type, orig_net, orig_addr, amount, dest_net, dest_addr, deposit_cnt, block_id, b.block_num, d.network_id, tx_hash, metadata, ready_for_claim, b.received_at, ready_time
-		FROM sync.deposit%[1]v as d INNER JOIN sync.block%[1]v as b ON d.network_id = b.network_id AND d.block_id = b.id
-		WHERE d.network_id = $1 AND ready_for_claim = true AND ready_time IS NOT NULL
-		ORDER BY d.deposit_cnt DESC LIMIT $2`, p.tableSuffix)
-
-	rows, err := p.getExecQuerier(dbTx).Query(ctx, getDepositsSQL, networkID, limit)
-	if err != nil {
-		return nil, err
-	}
-
-	deposits := make([]*etherman.Deposit, 0, len(rows.RawValues()))
-
-	for rows.Next() {
-		var (
-			deposit etherman.Deposit
-			amount  string
-		)
-		err = rows.Scan(&deposit.Id, &deposit.LeafType, &deposit.OriginalNetwork, &deposit.OriginalAddress, &amount, &deposit.DestinationNetwork, &deposit.DestinationAddress,
-			&deposit.DepositCount, &deposit.BlockID, &deposit.BlockNumber, &deposit.NetworkID, &deposit.TxHash, &deposit.Metadata, &deposit.ReadyForClaim, &deposit.Time, &deposit.ReadyTime)
-		if err != nil {
-			return nil, err
-		}
-		deposit.Amount, _ = new(big.Int).SetString(amount, 10) //nolint:gomnd
-		deposits = append(deposits, &deposit)
-	}
-
-	return deposits, nil
 }
