+++ test/e2e/edge_test.go
@@ -19,13 +19,11 @@ import (
 
 func depositFromL1(ctx context.Context, opsman *operations.Manager, t *testing.T) {
 	amount := new(big.Int).SetUint64(250000000000000000)
-	okbAddr := common.HexToAddress("0x82109a709138A2953C720D3d775168717b668ba6") // This means is okb
+	tokenAddr := common.Address{} // This means is eth
 	destAddr := common.HexToAddress("0xc949254d682d8c9ad5682521675b8f43b102aec4")
 	var destNetwork uint32 = 1
 	// L1 Deposit
-	err := opsman.ApproveERC20OKB(ctx, okbAddr, amount)
-	require.NoError(t, err)
-	err = opsman.SendL1Deposit(ctx, okbAddr, amount, destNetwork, &destAddr)
+	err := opsman.SendL1Deposit(ctx, tokenAddr, amount, destNetwork, &destAddr)
 	require.NoError(t, err)
 
 	deposits, err := opsman.GetBridgeInfoByDestAddr(ctx, &destAddr)
@@ -39,8 +37,8 @@ func depositFromL2(ctx context.Context, opsman *operations.Manager, t *testing.T
 	// Send L2 Deposit to withdraw the some funds
 	var destNetwork uint32 = 0
 	amount := new(big.Int).SetUint64(100000000000000000)
-	tokenAddr := common.Address{} // This means is okb
-	destAddr := common.HexToAddress("0x2ecf31ece36ccac2d3222a303b1409233ecbb225")
+	tokenAddr := common.Address{} // This means is eth
+	destAddr := common.HexToAddress("0x3C44CdDdB6a900fa2b585dd299e03d12FA4293BC")
 	err := opsman.SendL2Deposit(ctx, tokenAddr, amount, destNetwork, &destAddr)
 	require.NoError(t, err)
 
@@ -112,7 +110,7 @@ func TestEdgeCase(t *testing.T) {
 		require.NoError(t, opsman.StartBridge())
 		time.Sleep(st * time.Second)
 
-		depositFromL2(ctx, opsman, t)
-		depositFromL1(ctx, opsman, t)
+		depositFromL2X1(ctx, opsman, t)
+		depositFromL1X1(ctx, opsman, t)
 	})
 }
