+++ test/scripts/claim/main.go
@@ -65,7 +65,7 @@ func main() {
 		ExitRoots: []common.Hash{common.HexToHash(proof.MainExitRoot), common.HexToHash(proof.RollupExitRoot)},
 	}
 	log.Info("Sending claim tx...")
-	err = c.SendClaimAndWait(ctx, bridgeData, smt, globalExitRoot, auth)
+	err = c.SendClaim(ctx, bridgeData, smt, globalExitRoot, auth)
 	if err != nil {
 		log.Fatal("error: ", err)
 	}
