+++ utils/gerror/error.go
@@ -15,6 +15,8 @@ var (
 	ErrDepositNotSynced = errors.New("not synchronized deposit")
 	// ErrNetworkNotRegister is used when the networkID is not registered in the bridge
 	ErrNetworkNotRegister = errors.New("not registered network")
+
+	// X1
 	// ErrInternalErrorForRpcCall bridge web call service, when occur internal error(db error, redis error, network error...), return this error
 	ErrInternalErrorForRpcCall = errors.New("your request could not be processed, please try again later")
 )
